Rename handlers to cli handler Listen to ipv6 loopback address
